requirements.txt
streamlit
easyocr
opencv-python-headless
numpy
Pillow
torch
torchvision